requirements.txt
dynaconf
pydantic
pydantic_settings

aiogram
aiogram-dialog
cachetools

fluentogram
jinja2

gspread_asyncio

faststream
nats-py
msgpack

alembic
black
sqlalchemy
psycopg
psycopg-binary

rich

pytest
pytest-asyncio
